Temporarily disables XXE protection
In case the entity loader is disabled application wide, we still need to read this file so we need to disable it temporarily. 
Since we know this file can be trusted, there's no risk. Except if an exploit manages to replace this file, in which case this exploit is the least of our worries.